change it

+++ My_Class_App/Models/DomainModels/ClassItem.cs
@@ -8,6 +8,6 @@ namespace My_Classes_App.Models
 
 
         [JsonIgnore]
-        public double totalCredits => Class.NumberOfCredits;
+        public int totalCredits => (int)Class.NumberOfCredits;
     }
 }
