+++ My_Class_App/Models/DomainModels/IClass.cs
@@ -5,7 +5,7 @@ namespace My_Classes_App.Models
     public interface IClass
     {
         void Load(IRepository<Class> data);
-        double totalCredits { get; }
+        int totalCredits { get; }
         int? Count { get; }
         IEnumerable<ClassItem> List { get; }
         ClassItem GetById(int id);
