+++ My_Class_App/Models/DomainModels/MyClass.cs
@@ -52,7 +52,7 @@ namespace My_Classes_App.Models
             }
         }
 
-        public double totalCredits => items.Sum(i => i.totalCredits);
+        public int totalCredits => items.Sum(i => i.totalCredits);
         public int? Count => session.GetInt32(CountKey) ?? requestCookies.GetInt32(CountKey);
         public IEnumerable<ClassItem> List => items;
